+++ auto-file-and-function-comments/CHANGELOG.md
@@ -12,3 +12,7 @@ All notable changes to the "Auto File and Function Comments" extension will be d
 
 ## [1.0.2]
 - Fixed ability to change author name.
+
+## [1.0.3]
+- Added basic ability to comment C++ classes (still does not support inherited C++ classes).
+- Added ability to decide to include the date in the file comments and add additional comment lines for file, method, and adt comments.